has been fixed relations, name of method in transaction service, was deleted mappers, rename package from entity to model.

+++ src/main/java/com/org/house/repository/AutomatonRepository.java
@@ -1,9 +1,10 @@
 package com.org.house.repository;
 
-import com.org.house.entity.Automaton;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.org.house.model.Automaton;
+
 @Repository
-public interface AutomatonRepository extends JpaRepository<Automaton, Integer> {
+public interface AutomatonRepository extends JpaRepository<Automaton, Long> {
 }
