+++ src/main/java/com/org/house/repository/AutomatonStateRepository.java
@@ -1,11 +1,12 @@
 package com.org.house.repository;
 
-import com.org.house.entity.AutomatonState;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.org.house.model.AutomatonState;
 
 import java.util.Optional;
 
-public interface AutomatonStateRepository extends JpaRepository<AutomatonState, Integer> {
+public interface AutomatonStateRepository extends JpaRepository<AutomatonState, Long> {
 
-    Optional<AutomatonState> findByAutomatonId(int automatonId);
+    Optional<AutomatonState> findByAutomatonId(long automatonId);
 }
