+++ src/main/java/com/org/house/repository/TransactionRepository.java
@@ -1,13 +1,14 @@
 package com.org.house.repository;
 
-import com.org.house.entity.Transaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.org.house.model.Transaction;
 
 import java.util.Date;
 import java.util.List;
 
 @Repository
-public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
+public interface TransactionRepository extends JpaRepository<Transaction, Long> {
     List<Transaction> findByDateBetween(Date fromDate,Date dateBefore);
 }
